+++ vibrato/src/dictionary.rs
@@ -216,8 +216,8 @@ impl Dictionary {
     ///  - the set of new ids are not same as that of old ids.
     pub fn map_connection_ids_from_iter<L, R>(mut self, lmap: L, rmap: R) -> Result<Self>
     where
-        L: IntoIterator<Item = u16>,
-        R: IntoIterator<Item = u16>,
+        L: IntoIterator<Item = u32>,
+        R: IntoIterator<Item = u32>,
     {
         let mapper = ConnIdMapper::from_iter(lmap, rmap)?;
         self.data.system_lexicon.map_connection_ids(&mapper);
